requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inexactsearch==1.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
openai==1.37.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
preshed==3.0.9
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
silpa_common==0.3
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundex==1.1.3
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==3.0.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
